+++ src/NetStandard/E.Standard.WebGIS.Tools/Editing/Desktop/Advanced/DeleteSelectedSubsetOriginals.cs
@@ -171,7 +171,8 @@ public class DeleteSelectedSubsetOriginals : IApiServerToolAsync, IApiChildTool
             throw new ArgumentException("Can't find edit theme for selelction");
         }
 
-        if (!await editEnvironment.DeleteFeatures(editTheme, features))
+        var commitResult = await editEnvironment.DeleteFeatures(editTheme, features);
+        if (!commitResult)
         {
             throw new Exception("Löschen ist nicht möglich");
         }
@@ -188,7 +189,7 @@ public class DeleteSelectedSubsetOriginals : IApiServerToolAsync, IApiChildTool
                     Title = u.ToTitle(editTheme)
                 }).ToArray()
             /*, RefreshSelection = true*/
-        };
+        }.ApplyCommitFeatureResult(commitResult);
     }
 
     #endregion